211108_insidehome commit 8차 통합

+++ insidehome/src/main/java/home/inside/board/controller/BoardRefController.java
@@ -1,18 +1,16 @@
 package home.inside.board.controller;
 
+import java.util.HashMap;
+
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.multipart.MultipartHttpServletRequest;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import home.inside.board.service.IBoardService;
-import home.inside.board.util.ArticleMgrCommand;
-import home.inside.board.util.PageSearchCommand;
 
 @Controller
 @RequestMapping("/user/ref")
@@ -22,36 +20,43 @@ public class BoardRefController {
 
 	// 댓글등록 요청
 	@RequestMapping(value = "/regist.do", method = RequestMethod.POST)
-	public String insertRefSubmit(int boardNum, String content, HttpSession session, RedirectAttributes rttr) throws Exception {
+	public String insertRefSubmit(int boardNum, String content, HttpSession session, RedirectAttributes rttr)
+			throws Exception {
 		String nickname = (String) session.getAttribute("loginInside");
-//		if(nickname != null) {
-		nickname = "ggg";
-			System.out.println(boardNum+ nickname+ content);
+		rttr.addAttribute("boardNum", boardNum);
+		if (content != null && !content.trim().isEmpty()) {
 			ser.insertRef(boardNum, nickname, content);
-//		}
-			rttr.addAttribute("boardNum", boardNum);
+		}
 		return "redirect:/user/board/read.do";
 	}
 
 	// 댓글수정 요청
 	@RequestMapping(value = "/update.do", method = RequestMethod.POST)
-	public String updateRefSubmit(int num, int boardNum, String content, RedirectAttributes rttr) throws Exception {
-		if(content == null) {
-			rttr.addFlashAttribute("reFail", "fail");
-		}
-		System.out.println(num + content + boardNum);
-		ser.updateRef(num, content);
+	public String updateRefSubmit(int num, int boardNum, String content, RedirectAttributes rttr, HttpSession session)
+			throws Exception {
+		String nickname = (String) session.getAttribute("loginInside");
 		rttr.addAttribute("boardNum", boardNum);
-		rttr.addAttribute("read", "reE");
+		if (content != null && !content.trim().isEmpty()) {
+			HashMap<String, Object> tmpInfo = ser.isCheckWriterToUser(boardNum, num);
+			if (tmpInfo.get("REFWRITER").equals(nickname) || tmpInfo.get("BOARDWRITER").equals(nickname)) {
+				ser.updateRef(num, content);
+				rttr.addAttribute("read", "reE");
+			}
+		}
 		return "redirect:/user/board/read.do";
 	}
 
 	// 댓글삭제 요청
 	@RequestMapping(value = "/delete.do")
-	public String deleteRefSubmit(int num, int boardNum, RedirectAttributes rttr) throws Exception {
-		ser.deleteRef(num);
+	public String deleteRefSubmit(int num, int boardNum, RedirectAttributes rttr, HttpSession session)
+			throws Exception {
+		String nickname = (String) session.getAttribute("loginInside");
 		rttr.addAttribute("boardNum", boardNum);
-		rttr.addAttribute("read", "reD");
+		HashMap<String, Object> tmpInfo = ser.isCheckWriterToUser(boardNum, num);
+		if (tmpInfo.get("REFWRITER").equals(nickname) || tmpInfo.get("BOARDWRITER").equals(nickname)) {
+			ser.deleteRef(num);
+			rttr.addAttribute("read", "reD");
+		}
 		return "redirect:/user/board/read.do";
 	}
